Update Index.rst
Hi, 
This particular tutorial did more for me to break into how to use TypoScript and how to get it to work with the HTML template. I am really going to work a lot on it, and get it gramatically cleaned up and suggest quite a few more ideas. What would be better, for me to use this document as bibliography and create a new tutorial, or is it a possibility that I add in more comments and explanations in between for those new to the task?
There are a lot of changes in this page, but I believe there will be a little more clarity. I have tried maintaining consistency in the use of verbs such as the last list of possibilites. They were not all beginning with gerunds (e.g. using, building, etc.), so I reworked the phrases for that reason. 
I would like to have more explanations for every detail, but, by the time I get to the end, I find that I don´t remember all the details.
Please feel free to question me.
Tim